Updating Coroutines Live Data

+++ app/src/main/java/com/example/jetpackcomponentsapp/MainViewModel.kt
@@ -3,23 +3,26 @@ package com.example.jetpackcomponentsapp
 import android.app.Application
 import android.util.Log
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.AndroidViewModel
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.*
 import com.example.jetpackcomponentsapp.model.CustomModel
-import com.example.jetpackcomponentsapp.repository.CustomRepository
 import com.example.jetpackcomponentsapp.util.ConvertList
+import com.example.jetpackcomponentsapp.repository.CustomRepository
+import com.example.jetpackcomponentsapp.room.CustomEntity
 import com.example.jetpackcomponentsapp.util.Coroutines
 import kotlinx.coroutines.delay
 
-class MainViewModel : AndroidViewModel {
+public class MainViewModel : AndroidViewModel {
 
-    private var customRepository : CustomRepository
+    companion object {
+        private val TAG = MainViewModel::class.java.getSimpleName()
+    }
+
+    private val customRepository : CustomRepository
     private var liveUpdate : MutableLiveData<CustomModel>
     private var liveStandBy : MutableLiveData<Boolean>
 
     constructor(application: Application) : super(application) {
-        customRepository = CustomRepository.getInstance(application)
+        customRepository = CustomRepository(application)
         liveUpdate = MutableLiveData(CustomModel())
         liveStandBy = MutableLiveData(null)
     }
@@ -40,29 +43,15 @@ class MainViewModel : AndroidViewModel {
         liveStandBy.postValue(false)
     }
 
-    public fun getLoadState() : LiveData<Boolean> {
+    public fun observeLoadState() : LiveData<Boolean> {
         return liveStandBy
     }
 
-    @Deprecated("For Static Data")
-    public fun setItems() { Coroutines.io (this@MainViewModel, {
-        liveStandBy.postValue(true)
-        customRepository.deleteAll()
-        customRepository.insert(ConvertList.toEntity(CustomModel(R.drawable.ic_launcher_foreground, "name 0")))
-        customRepository.insert(ConvertList.toEntity(CustomModel(R.drawable.ic_launcher_foreground, "name 1")))
-        customRepository.insert(ConvertList.toEntity(CustomModel(R.drawable.ic_launcher_foreground, "name 2")))
-        customRepository.insert(ConvertList.toEntity(CustomModel(R.drawable.ic_launcher_foreground, "name 3")))
-        customRepository.insert(ConvertList.toEntity(CustomModel(R.drawable.ic_launcher_foreground, "name 4")))
-        customRepository.insert(ConvertList.toEntity(CustomModel(R.drawable.ic_launcher_foreground, "name 5")))
-        customRepository.insert(ConvertList.toEntity(CustomModel(R.drawable.ic_launcher_foreground, "name 6")))
-        customRepository.insert(ConvertList.toEntity(CustomModel(R.drawable.ic_launcher_foreground, "name 7")))
-        customRepository.insert(ConvertList.toEntity(CustomModel(R.drawable.ic_launcher_foreground, "name 8")))
-        viewWillAppear()
-    } ) }
-
-    public fun setUpdate(item : CustomModel) { Coroutines.io(this@MainViewModel, {
-        liveStandBy.postValue(true)
-        liveUpdate.postValue(item)
+    public fun setUpdate(item : CustomModel?) { Coroutines.io(this@MainViewModel, {
+        if (item != null) {
+            liveStandBy.postValue(true)
+            liveUpdate.postValue(item)
+        }
     } ) }
 
     public fun observeUpdate() : LiveData<CustomModel> {
@@ -72,25 +61,26 @@ class MainViewModel : AndroidViewModel {
     public fun insertItem(item : CustomModel) { Coroutines.io(this@MainViewModel, {
         liveStandBy.postValue(true)
         customRepository.insert(
-                ConvertList.toEntity(item)
+            ConvertList.toEntity(item)
         )
         viewWillAppear()
     } ) }
 
     public fun updateItem(updated : String) { Coroutines.io(this@MainViewModel, {
         liveStandBy.postValue(true)
-        liveUpdate.getValue()?.name = updated
+        val old : CustomModel? = liveUpdate.value
         customRepository.update(
-            ConvertList.toEntity(liveUpdate.value ?: CustomModel())
+            ConvertList.toEntity(CustomModel(old?.id, updated, old?.icon))
         )
         viewWillAppear()
     }) }
 
-    public fun deleteItem(item : CustomModel) { Coroutines.io(this@MainViewModel, {
+    public fun deleteItem(item : CustomModel?) { Coroutines.io(this@MainViewModel, {
         liveStandBy.postValue(true)
-        customRepository.delete(
+        if (item != null)
+            customRepository.delete(
                 ConvertList.toEntity(item)
-        )
+            )
         viewWillAppear()
     } ) }
 
@@ -100,13 +90,16 @@ class MainViewModel : AndroidViewModel {
         viewWillAppear()
     }) }
 
-    public fun getItems() : LiveData<List<CustomModel>> {
+    public fun observeItems() : LiveData<List<CustomModel>> {
         return ConvertList.toLiveDataListModel(
-            customRepository.getAll()
+            customRepository.getAll() ?: MutableLiveData<List<CustomEntity>>(listOf<CustomEntity>())
         )
     }
 
     override fun onCleared() {
+        Coroutines.io(this@MainViewModel, {
+            customRepository.onCLose()
+        })
         super.onCleared()
     }
 }
